added usecase to provide the data for DashboardPresenter.kt
