+++ app/src/main/java/com/sample/experiments/ui/DashboardPresenter.kt
@@ -1,121 +1,62 @@
 package com.sample.experiments.ui
 
-import com.sample.experiments.R
 import com.sample.experiments.domain.DashboardItem
-import com.sample.experiments.domain.GetItemsUseCase
+import com.sample.experiments.domain.GetItemsUseCase2
 import com.sample.experiments.domain.TimeProvider
 import com.sample.experiments.ui.items.BasePresenter
 import com.sample.experiments.ui.items.timer.CopyPerRowUIModel
-import com.sample.experiments.ui.items.timer.OnePerModelUIModel
 import com.sample.experiments.ui.items.timer.SingleEngineTimerUIModel
 import io.reactivex.Observable
 import io.reactivex.android.schedulers.AndroidSchedulers
-import java.text.SimpleDateFormat
-import java.util.*
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
-import kotlin.collections.ArrayList
-import kotlin.random.Random
 
-
-class DashboardPresenter
-@Inject constructor(
-    private val getItemsUseCase: GetItemsUseCase,
+class DashboardPresenter @Inject constructor(
+    private val getItemsUseCase: GetItemsUseCase2,
     private val timeProvider: TimeProvider
 ) : BasePresenter<DashBoardView>() {
 
-    private val formatter = SimpleDateFormat("HH:mm:ss", Locale.ENGLISH)
-
-
-//    fun loadItems() {
-//        val items = getItemsUseCase()
-//        view?.showItems(items)
-//    }
-
 
     fun loadItems() {
-//        compositeDisposable.add(Observable.interval(
-//            1,
-//            TimeUnit.SECONDS,
-//            AndroidSchedulers.mainThread()
-//        )
-//            .subscribe {
-//                items.forEach {
-//                    if (it is SingleEngineTimerModel) {
-//                        it.tick()
-//                    }
-//                }
-//            })
-        compositeDisposable.add(Observable.just(timerItems)
+        compositeDisposable.add(getItemsUseCase
+            .invoke()
             .publish {
-                Observable.merge(it.take(1), it.updateItems())
+                Observable.merge(
+                    //pass on the first items we receive
+                    it.take(1),
+                    //updates on every second
+                    it.updateItems()
+                )
             }
+            .observeOn(AndroidSchedulers.mainThread())
             .subscribe {
                 view?.showItems(it)
             })
     }
 
-    private val timerItems: List<DashboardItem> by lazy {
-        val baseEndTime = System.currentTimeMillis() + 30_000 //30 seconds from now
-        val items = ArrayList<DashboardItem>()
-        repeat(1000) {
-            val date = Date(baseEndTime + Random.nextLong(90_000))
-            when (Random.nextInt(1, 5)) {
-                1 -> {
-                    items.add(SingleEngineTimerUIModel(date.time, date, timeProvider))
-                }
-                2 -> {
-                    items.add(OnePerModelUIModel(date, timeProvider))
-                }
-                3 -> {
-                    val remaining = getRemainingTime(date.time, timeProvider.currentTime())
-                    items.add(
-                        CopyPerRowUIModel(
-                            date,
-                            "TAM= Ends at: " + formatter.format(date),
-                            getRemainingTimeMessage(remaining),
-                            getFinishedColor(remaining)
-                        )
-                    )
-                }
-            }
 
-        }
-        items
-    }
-
-    private fun getRemainingTime(endDate: Long, current: Long) =
-        endDate - current
-
-    private fun getRemainingTimeMessage(remaining: Long) =
-        if (remaining > 0) {
-            (remaining / 1_000).toInt().toString() + " seconds"
-        } else {
-            "Finish"
-        }
-
-    private fun getFinishedColor(remainingTime : Long) =
-        if (remainingTime > 0) R.color.red else R.color.green
-
-    private fun Observable<List<DashboardItem>>.updateItems() : Observable<List<DashboardItem>> =
+    private fun Observable<List<DashboardItem>>.updateItems(): Observable<List<DashboardItem>> =
         switchMap { items ->
             Observable.interval(
                 1,
                 TimeUnit.SECONDS,
                 AndroidSchedulers.mainThread()
             ).doOnNext {
+                //updating single engines without changing their objects
+                // view is already subscribed to the change events
                 items.forEach {
                     if (it is SingleEngineTimerUIModel) {
                         it.tick()
                     }
                 }
             }
+                //copy and create new UIModel object for `CopyPerRowUIModel` and send it downstream
                 .scan(items, { t1, t2 ->
                     t1.map {
                         val current = timeProvider.currentTime()
-                        if(it is CopyPerRowUIModel){
+                        if (it is CopyPerRowUIModel) {
                             it.copyWithCurrentTime(current)
-                        }else{
+                        } else {
                             it
                         }
                     }
